refactor : 댓글 api 수정에 따른 댓글 컨테이너 리펙토링, 시간 파싱 함수 및 시간 표시 함수 수정
