+++ public/mock/studyDetailQustionMockData.json
@@ -1,62 +1,54 @@
 {
   "contents": [
     {
-      "commentId": 1,
       "questionId": 1,
       "parentId": null,
-      "profileImageUrl": "",
+      "profileImageUrl": "http://prgrms.stuti/image1.jpg",
       "memberId": 1,
       "nickname": "팽",
       "contents": "글이 좋네요 어떤스터디에 참여하고 계신가요?",
-      "createdAt": "2000.02.22 10:00:00",
+      "createdAt": "2000-02-22 10:00:00",
       "children": [
         {
           "parentId": 1,
-          "commentId": 2,
           "questionId": 4,
           "profileImageUrl": "",
           "memberId": 2,
           "nickname": "키아",
           "contents": "김팽박이라는 스터디에 참여하고있습니다.",
-          "createdAt": "2000.02.22 10:00:00"
+          "createdAt": "2000-02-22 10:00:00"
         },
         {
           "parentId": 1,
-          "commentId": 3,
           "questionId": 5,
-          "profileImageUrl": "",
+          "profileImageUrl": "http://prgrms.stuti/image33.jpg",
           "memberId": 1,
           "nickname": "팽",
           "contents": "아직 모집중인가요? 저도 참여하고싶습니다.",
-          "createdAt": "2000.02.22 10:00:00"
+          "createdAt": "2000-02-22 10:00:00"
         },
         {
           "parentId": 1,
-          "commentId": 4,
           "questionId": 6,
-          "profileImageUrl": "",
+          "profileImageUrl": "http://prgrms.stuti/image33.jpg",
           "memberId": 2,
           "nickname": "키아",
           "contents": "네~ 스터디 검색해서 신청해주세요 ~",
-          "createdAt": "2000.02.22 10:00:00"
+          "createdAt": "2000-02-22 10:00:00"
         }
       ]
     },
     {
-      "questionId": 2,
+      "questionId": 10,
       "parentId": null,
-      "profileImageUrl": "",
-      "memberId": 3,
-      "nickname": "프룽이2",
-      "content": "참여하고 싶어요!",
-      "createdAt": "2022.02.22 10:00",
+      "profileImageUrl": "http://prgrms.stuti/image1.jpg",
+      "memberId": 1,
+      "nickname": "팽",
+      "contents": "글이 좋네요 어떤스터디에 참여하고 계신가요?",
+      "createdAt": "2022-08-03 16:42:00",
       "children": []
     }
   ],
-  "page": 1,
-  "size": 5,
-  "totalPages": 10,
-  "totalElements": 50,
-  "isFirst": true,
-  "isLast": false
+  "isLast": false,
+  "totalElements": 50
 }
